+++ lib/ResqueScheduler/Job/Status.php
@@ -8,7 +8,7 @@
  * @copyright   Copyright 2013, Wan Qi Chen <kami@kamisama.me>
  * @license     MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace Kamisama\ResqueScheduler\Job;
+namespace ResqueScheduler\Job;
 
 class Status extends \Resque_Job_Status
 {